streaming-data-with-pubsub-and-dataflow/requirements.txt
black
pylint
google-cloud-pubsub
wheel
apache-beam[gcp]
